Refactor to prepare for custom log position provider
